requirements.txt
flask>=2.0.0
numpy>=1.21.0
faiss-cpu>=1.7.0
requests>=2.25.0
python-dotenv>=0.19.0
anthropic>=0.7.0
scikit-learn>=1.0.0
nltk>=3.8